rename: recommend 컴포넌트 discover로 이동
